// src/types/database.types.tsx
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      cmr_reserves: {
        Row: {
          cmr_id: string
          comment: string | null
          created_at: string
          id: string
          photo_storage_path: string | null
          reserve_type: string
          side: string
          user_id: string
        }
        Insert: {
          cmr_id: string
          comment?: string | null
          created_at?: string
          id?: string
          photo_storage_path?: string | null
          reserve_type: string
          side: string
          user_id: string
        }
        Update: {
          cmr_id?: string
          comment?: string | null
          created_at?: string
          id?: string
          photo_storage_path?: string | null
          reserve_type?: string
          side?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "cmr_reserves_cmr_id_fkey"
            columns: ["cmr_id"]
            isOneToOne: false
            referencedRelation: "cmr_documents"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}